pyproject.toml
[tool.poetry]
name = "dip"
version = "0.1.0"
description = ""
authors = ["nomicode <23469+nomicode@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "dip", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pillow = "^10.4.0"
numpy = "<2.0"
opencv-python = "^4.10.0.84"
scikit-image = "^0.24.0"
scikit-learn = "^1.5.2"
segment-anything = "^1.0"
psd-tools = "^1.10.1"
controlnet-aux = "^0.0.9"
imagehash = "^4.3.1"
mediapipe = "^0.10.15"
rembg = "^2.0.59"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
flake8 = "^6.1.0"
black = "^23.7.0"
autoflake = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli_segment_anything = "scripts.cli_segment_anything:main"
controlnet = "scripts.controlnet:main"
img_quant = "scripts.img_quant:main"
imghash = "scripts.imghash:main"
imgproc = "scripts.imgproc:main"

[tool.black]
line-length = 79
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''
